moved spinner into internal package
