+++ internal/spinner/spinner.go
@@ -1,28 +1,32 @@
-package main
+package spinner
 
 import (
 	"fmt"
 	"time"
 )
 
-func withSpinner(msg string, task func(finished chan<- struct{})) {
-	// start task
-	finished := make(chan struct{})
-	go task(finished)
+// While executes f, displays an animated spinner while f runs, and stops when f returns.
+func While(f func()) {
+	c := make(chan struct{})
 
-	fmt.Print(msg)
+	go func(c chan<- struct{}) {
+		defer close(c)
+		f()
+	}(c)
 
+	Until(c)
+}
+
+// Until displays an animated spinner until reading from c succeeds. It is recommended to simply close c to achieve this.
+func Until(c <-chan struct{}) {
 	// start spinning animation
 	stop := make(chan struct{})
 	go spin(stop)
 
 	// wait for task to finish, then tell spinner to stop
-	stop <- <-finished
+	stop <- <-c
 	// wait for spinner to stop
 	<-stop
-
-	fmt.Println()
-	return
 }
 
 // spin animates a spinner until it receives something on the stop channel. It then clears the spinning character and closes the stop channel, signaling that it's done.
